+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Footer from './components/Footer'
 import Header from './components/Header/Header'
-import MountainComparer from './components/MountainComparer'
-import MyVirtualCookbook from './components/MyVirtualCookbook'
+import MountainComparer from './components/Portfolio/MountainComparer'
+import MyVirtualCookbook from './components/Portfolio/MyVirtualCookbook'
 import ResumeLink from './components/ResumeLink'
 import Section from './components/Section'
 import Skills from './components/Skills'
